requirements.txt
accelerate==0.27.2
aiohttp==3.10.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
exceptiongroup==1.2.0
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gymnasium==0.29.0
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
miniwob==1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
numpy==1.24.1
openai==0.28.0
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
selenium==4.16.0
setuptools==58.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.13.3
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.33.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.0.141
urllib3==2.2.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0